BBTSP

+++ CAAIScpp/src/utilities/RandomNumber.cpp
@@ -22,7 +22,7 @@ int RandRange(int low, int high)
 //Generate one random real in the range [0, 1) 
 double RandDbl()
 {
-	return (double)rand()/(RAND_MAX+1);
+	return (double)rand()/((long)RAND_MAX + 1);
 }
 //Set seed for rand()
 void SetRandSeed(int seed)
